[#13] Use any up-rust upstreamed code
* Remove transport handle, collapse down to lib and transport
* Rearrange transport engine
* Implement LocalUriProvider
* Make use of new up_rust::UUri::try_from_parts
* Show usage of uP-L2 APIs in hello_client and hello_service examples
  * Fix logic of Response handling to mean no commstatus is OK

Implements [#6], [#8]
