+++ up-transport-vsomeip/src/storage/message_handler_registry.rs
@@ -330,10 +330,10 @@ impl InMemoryMessageHandlerRegistry {
             message_handler_id_and_listener_config.deref_mut(),
             listener_config,
         )
-        .ok_or(
-            // TODO: Update this to include listener_config when we update to up-rust on crates.io with ComparableListener Debug impl
-            UStatus::fail_with_code(UCode::NOT_FOUND, "No listener_id for listener_config"), // UStatus::fail_with_code(UCode::NOT_FOUND, format!("No listener_id for listener_config: {listener_config:?}"))
-        )?;
+        .ok_or(UStatus::fail_with_code(
+            UCode::NOT_FOUND,
+            "No listener_id for listener_config",
+        ))?;
 
         if let Err(warn) =
             Self::free_message_handler_id(free_message_handler_ids.deref_mut(), message_handler_id)
