+++ up-transport-vsomeip/tests/publisher_subscriber.rs
@@ -16,7 +16,7 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::time::Instant;
-use up_rust::{UListener, UMessage, UMessageBuilder, UPayloadFormat, UStatus, UTransport, UUri};
+use up_rust::{UListener, UMessage, UMessageBuilder, UPayloadFormat, UTransport, UUri};
 use up_transport_vsomeip::UPTransportVsomeip;
 
 const TEST_DURATION: u64 = 1000;
@@ -52,10 +52,6 @@ impl UListener for SubscriberListener {
 
         info!("We received payload_string: {payload_string}");
     }
-
-    async fn on_error(&self, err: UStatus) {
-        trace!("{:?}", err);
-    }
 }
 
 #[tokio::test(flavor = "multi_thread")]
@@ -69,20 +65,13 @@ async fn publisher_subscriber() {
     let ue_version_major = 1;
     let resource_id = 0x8001;
 
-    let publisher_topic = UUri {
-        authority_name: authority_name.to_string(),
-        ue_id: ue_id as u32,
-        ue_version_major,
-        resource_id,
-        ..Default::default()
-    };
+    let publisher_topic =
+        UUri::try_from_parts(authority_name, ue_id, ue_version_major, resource_id).unwrap();
 
-    let subscriber_res = UPTransportVsomeip::new(
-        &authority_name.to_string(),
-        &"me_authority".to_string(),
-        subscriber_ue_id,
-        None,
-    );
+    let subscriber_uri =
+        UUri::try_from_parts(authority_name, subscriber_ue_id, ue_version_major, 0).unwrap();
+
+    let subscriber_res = UPTransportVsomeip::new(subscriber_uri, &"me_authority".to_string(), None);
 
     let Ok(subscriber) = subscriber_res else {
         panic!("Unable to establish subscriber");
@@ -103,12 +92,8 @@ async fn publisher_subscriber() {
 
     tokio::time::sleep(Duration::from_millis(1000)).await;
 
-    let publisher_res = UPTransportVsomeip::new(
-        &authority_name.to_string(),
-        &"me_authority".to_string(),
-        ue_id,
-        None,
-    );
+    let publisher_uri = UUri::try_from_parts(authority_name, ue_id, 1, 0).unwrap();
+    let publisher_res = UPTransportVsomeip::new(publisher_uri, &"me_authority".to_string(), None);
 
     let Ok(publisher) = publisher_res else {
         panic!("Unable to establish publisher");
